requirements-test.txt
fastapi
uvicorn[standard]
llama-cpp-python
pytest
pytest-mock
httpx